requirements.txt
blinker==1.4
click==7.1.2
Flask==1.1.4
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
Flask-QRcode==3.0.0
flask-selfdoc==1.3.0
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.28
mysqlclient==2.1.0
Pillow==9.0.1
protobuf==3.19.4
PyYAML==6.0
qrcode==7.3.1
Werkzeug==1.0.1
